Add files via upload
2.Write a multithreaded program that calculates various statistical values for a list of numbers. This program will be passed a series of numbers on the command line and will then create three separate worker threads. One thread will determine the average of the numbers, the second will determine the maximum value, and the third will determine the minimum value. For example, suppose your program is passed the integers 
90 81 78 95 79 72 85 
The program will report 
The average value is 82 
The minimum value is 72 
The maximum value is 95 
The variables representing the average, minimum, and maximum values will be stored globally. The worker threads will set these values, and the parent thread will output the values once the workers have exited .

Description:
Multithreading:
Thread is flow of execution through the process of code. A thread is also known as light weigth process.
Thread provides to improve application performance throught the parallelism.
Now days many of the people writing  in a single thread programming. Threading is nothing but a process. Multithreading is nothing but a multiple process running parallel.
An operating system which runs multiple threads of single application is called multithreading operating system.
Multithreading it is a concept of parallel system in operating system where there are too many processors working on different threads each threads having dedicated task.
In programming , multithreading allows a program to run multiple tasks concurrently which is very useful for heavy data processing  or tasks which rely on external factors ,such as slower storage or internet services. 
Advantages:
The main advantage in multithreading is proper utilization or maximum utilization is possible.
It decreases the cost of maintenance.
In the multithreading it is time saving.
In the multithreading it improves the performance of complex applications.
It is having parallelize tasks.
Disadvantages:
It is having unpredictable results.
Increased difficulty level in writing a program.
Complex debugging and testing processes.